Improved TeamVesting to manage everything and fixed InvestorVesting
